refactor(storage): use explicit control flow in rules-evaluator and mirror test path
Replace multi-line wrapping ternaries in normalizeAuth and buildRequestObject
with explicit if/else control flow per Section 3b & 3c of docs/code-conventions.md.
Revert unrelated edit to test/rules/parity/harness.ts and move Storage rules
regression tests to mirrored path test/storage/sandbox/rules-evaluator.test.ts.

Conformance-Exempt: code style, test mirroring, and scope cleanup
